+++ src-ui/app/components/dashboard-navbar/dashboard-navbar.component.ts
@@ -1,18 +1,12 @@
 import { Component, DestroyRef, OnInit } from '@angular/core';
-import { asyncScheduler, combineLatest, map, Observable, startWith, throttleTime } from 'rxjs';
-import { LighthouseConsoleService } from 'src-ui/app/services/lighthouse-console.service';
-import { GpuAutomationsService } from '../../services/gpu-automations.service';
+import { map, Observable } from 'rxjs';
 import { fade } from '../../utils/animations';
-import { NvmlService } from '../../services/nvml.service';
-import { ElevatedSidecarService } from '../../services/elevated-sidecar.service';
 import { Router } from '@angular/router';
 import { animate, state, style, transition, trigger } from '@angular/animations';
 import { BackgroundService } from '../../services/background.service';
-import { OscService } from '../../services/osc.service';
 import { BrightnessCctAutomationService } from '../../services/brightness-cct-automation.service';
 import { ModalService } from 'src-ui/app/services/modal.service';
 import { DeveloperDebugModalComponent } from '../developer-debug-modal/developer-debug-modal.component';
-import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { UpdateService } from 'src-ui/app/services/update.service';
 
 function slideMenu(name = 'slideMenu', length = '.2s ease', root = true) {
@@ -131,18 +125,10 @@ type SubMenu = 'GENERAL' | 'VRCHAT' | 'HARDWARE' | 'MISCELLANEOUS' | 'SETTINGS';
   standalone: false,
 })
 export class DashboardNavbarComponent implements OnInit {
-  settingErrors: Observable<boolean>;
-  gpuAutomationsErrors: Observable<boolean>;
-  lighthouseConsoleErrors: Observable<boolean>;
   subMenu: SubMenu = 'GENERAL';
   updateAvailable: Observable<boolean>;
 
   constructor(
-    private lighthouse: LighthouseConsoleService,
-    private gpuAutomations: GpuAutomationsService,
-    private nvml: NvmlService,
-    private sidecar: ElevatedSidecarService,
-    private osc: OscService,
     private updateService: UpdateService,
     protected router: Router,
     protected background: BackgroundService,
@@ -151,63 +137,6 @@ export class DashboardNavbarComponent implements OnInit {
     private destroyRef: DestroyRef
   ) {
     this.updateAvailable = this.updateService.updateAvailable.pipe(map((a) => !!a.update));
-    this.lighthouseConsoleErrors = this.lighthouse.consoleStatus.pipe(
-      takeUntilDestroyed(this.destroyRef),
-      map((status) => !['UNKNOWN', 'SUCCESS', 'CHECKING'].includes(status)),
-      startWith(false)
-    );
-    this.settingErrors = combineLatest([this.lighthouseConsoleErrors]).pipe(
-      map((errorAreas: boolean[]) => !!errorAreas.find((a) => a))
-    );
-
-    this.gpuAutomationsErrors = combineLatest([
-      this.gpuAutomations.isEnabled(),
-      this.nvml.status,
-      this.sidecar.sidecarStarted,
-      this.gpuAutomations.msiAfterburnerStatus,
-      this.gpuAutomations.msiAfterburnerConfig,
-    ]).pipe(
-      throttleTime(300, asyncScheduler, { trailing: true, leading: true }),
-      map(
-        ([
-          gpuAutomationsEnabled,
-          nvmlStatus,
-          sidecarRunning,
-          msiAfterburnerStatus,
-          msiAfterburnerConfig,
-        ]) => {
-          // Global
-          if (!gpuAutomationsEnabled) return false;
-          if (!sidecarRunning) return true;
-          // Nvml
-          if (
-            [
-              'ELEVATION_SIDECAR_INACTIVE',
-              'NO_PERMISSION',
-              'Nvml_UNKNOWN_ERROR',
-              'UNKNOWN_ERROR',
-            ].includes(nvmlStatus)
-          )
-            return true;
-          // Afterburner
-          if (
-            (msiAfterburnerConfig.onSleepDisableProfile > 0 ||
-              msiAfterburnerConfig.onSleepEnableProfile > 0) &&
-            [
-              'NOT_FOUND',
-              'INVALID_EXECUTABLE',
-              'PERMISSION_DENIED',
-              'INVALID_FILENAME',
-              'INVALID_SIGNATURE',
-              'UNKNOWN_ERROR',
-            ].includes(msiAfterburnerStatus)
-          )
-            return true;
-          // No errors found
-          return false;
-        }
-      )
-    );
   }
 
   async ngOnInit(): Promise<void> {}