+++ CHANGELOG.md
@@ -12,16 +12,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GPU acceleration for SteamVR overlays (Community contribution by [BenjaminZehowlt](https://github.com/BenjaminZehowlt))
 - VR Headset device & battery level to the device list
 - Sleep mode toggle and sleep preparation to system tray menu
+- Message center for automatically diagnosing common issues
+- Exposed headset battery level and charging indicator over MQTT
 
 ### Changed
 
 - Added continuous scanning for lighthouses to replace manual scanning. 
 - Migrated to Tauri v2, and upgraded various dependencies.
 - Localized system tray menu
+- Updated theming
 
 ### Fixed
 
 - "Activate when you've been left alone" trigger for the shutdown sequence not triggering correctly.  
+- Battery and charging levels exposed over MQTT becoming unavailable randomly
 
 ## [1.14.7]